+++ TombEngine/Objects/TR4/Entity/tr4_knight_templar.cpp
@@ -195,7 +195,7 @@ namespace TEN::Entities::TR4
 					{
 						if (abs(pos.x - mesh.pos.Position.x) < BLOCK(1) &&
 							abs(pos.z - mesh.pos.Position.z) < BLOCK(1) &&
-							StaticObjects[mesh.staticNumber].shatterType == ShatterType::None)
+							GetStaticObject(mesh.staticNumber).shatterType == ShatterType::None)
 						{
 							ShatterObject(nullptr, &mesh, -64, LaraItem->RoomNumber, 0);
 							SoundEffect(SFX_TR4_SMASH_ROCK, &item->Pose);